refactor init_weights
