+++ src/fairchem/core/models/equiformer_v2/equiformer_v2.py
@@ -1,8 +1,14 @@
+"""
+Copyright (c) Meta, Inc. and its affiliates.
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+"""
+
 from __future__ import annotations
 
 import contextlib
 import logging
-import math
 import typing
 from functools import partial
 
@@ -30,7 +36,6 @@ from .layer_norm import (
     get_normalization_layer,
 )
 from .module_list import ModuleListInfo
-from .radial_function import RadialFunction
 from .so3 import (
     CoefficientMappingModule,
     SO3_Embedding,
@@ -41,6 +46,7 @@ from .so3 import (
 from .transformer_block import (
     TransBlockV2,
 )
+from .weight_initialization import eqv2_init_weights
 
 with contextlib.suppress(ImportError):
     pass
@@ -75,28 +81,6 @@ class EquiformerV2EnergyHead(EqV2ScalarHead):
             "equiformerV2_energy_head (EquiformerV2EnergyHead) class is deprecated in favor of equiformerV2_scalar_head  (EqV2ScalarHead)"
         )
         super().__init__(backbone, reduce=reduce)
-
-
-def eqv2_init_weights(m, weight_init):
-    if isinstance(m, (torch.nn.Linear, SO3_LinearV2)):
-        if m.bias is not None:
-            torch.nn.init.constant_(m.bias, 0)
-        if weight_init == "normal":
-            std = 1 / math.sqrt(m.in_features)
-            torch.nn.init.normal_(m.weight, 0, std)
-    elif isinstance(m, torch.nn.LayerNorm):
-        torch.nn.init.constant_(m.bias, 0)
-        torch.nn.init.constant_(m.weight, 1.0)
-    elif isinstance(m, RadialFunction):
-        m.apply(eqv2_uniform_init_linear_weights)
-
-
-def eqv2_uniform_init_linear_weights(m):
-    if isinstance(m, torch.nn.Linear):
-        if m.bias is not None:
-            torch.nn.init.constant_(m.bias, 0)
-        std = 1 / math.sqrt(m.in_features)
-        torch.nn.init.uniform_(m.weight, -std, std)
 
 
 @registry.register_model("equiformer_v2_backbone")
